Place photo_base class in a single file
