+++ cloelike/EuclidLikelihood_photo_2pcf.py
@@ -1,114 +1,11 @@
 # file: euclid_likelihoods.py
 import numpy as np
-from functools import lru_cache
 from cloelib.observables.photo import ShearTracer, PositionsTracer
 from cloelib.summary_statistics.angular_two_point import AngularTwoPoint
 from cloelib.summary_statistics.angular_correlation_function_wigner import (
     AngularCorrelationFunctionWigner,
 )
-
-
-class PhotoLikelihoodBase:
-    """
-    Base class for photometric likelihood calculations using angular correlation functions.
-    This class provides methods for preparing, binning, and masking data, as well as computing
-    likelihoods based on theoretical predictions and observed data vectors.
-    Args:
-        data (dict): Dictionary containing observational data, including '2pcf', 'theta', 'z_arr', and 'cov'.
-        settings (dict): Configuration settings, including 'scale_cuts'.
-        Background: Object representing background cosmology.
-        LinPerturbations: Object representing linear perturbations.
-        NonLinPerturbations: Object representing non-linear perturbations.
-        mode (str, optional): Mode of operation, default is "coupled".
-    Attributes:
-        data (dict): Observational data.
-        settings (dict): Configuration settings.
-        derived (dict): Dictionary for storing derived quantities.
-        Background: Background cosmology object.
-        LinPerturbations: Linear perturbations object.
-        NonLinPerturbations: Non-linear perturbations object.
-        scale_cuts: Scale cuts from settings.
-        zs: Redshift array from data.
-        mixmat: Mixing matrix, possibly rebinned.
-        weight_mat: Weight matrix used for binning.
-        masking_vector: Boolean mask for selecting data vector elements.
-    Methods:
-        _masking(arr, interval):
-            Returns a boolean mask for elements within the specified interval.
-        get_covariance_matrix_full():
-            Returns the full covariance matrix from the data.
-        get_data_vector_masked():
-            Returns the masked data vector.
-        get_covariance_matrix_masked_inv():
-            Returns the inverse of the masked covariance matrix.
-        get_theory_vector_full(parameters):
-            Returns the full theoretical prediction vector for given parameters.
-        get_theory_vector_masked(parameters):
-            Returns the masked theoretical prediction vector for given parameters.
-        loglike(parameters):
-            Computes the log-likelihood for the given parameters using the masked data and theory vectors.
-    """
-
-    def __init__(
-        self,
-        data,
-        settings,
-        Background,
-        LinPerturbations,
-        NonLinPerturbations,
-        ells_integration=np.arange(2, 60000),
-    ):
-        self.data = data
-        self.settings = settings
-        self.derived = {}
-        self.Background = Background
-        self.LinPerturbations = LinPerturbations
-        self.NonLinPerturbations = NonLinPerturbations
-        self.ells_integration = ells_integration
-        self.scale_cuts = settings["scale_cuts"]
-        self.zs = data["z_arr"]
-
-    def _masking(self, arr, interval):
-        return (arr >= interval[0]) & (arr <= interval[1])
-
-    def get_masking_vector(self):
-        return np.array([], dtype=bool)
-
-    def get_covariance_matrix_full(self):
-        return self.data["cov"]
-
-    def get_data_vector_full(self):
-        return np.array([])
-
-    @lru_cache(maxsize=None)
-    def get_masking_vector_cached(self):
-        """Compute (once) and cache the combined boolean mask."""
-        return self.get_masking_vector()
-
-    @lru_cache(maxsize=None)
-    def get_data_vector_masked(self):
-        mask = self.get_masking_vector_cached()
-        return self.get_data_vector_full()[mask]
-
-    @lru_cache(maxsize=None)
-    def get_covariance_matrix_masked_inv(self):
-        """Compute (once) and cache inverse masked covariance matrix."""
-        cov = self.get_covariance_matrix_full()
-        mask = self.get_masking_vector_cached()
-        cov_masked = cov[mask][:, mask]
-        return np.linalg.inv(cov_masked)
-
-    def get_theory_vector_full(self, parameters):
-        return np.array([])
-
-    def get_theory_vector_masked(self, parameters):
-        mask = self.get_masking_vector_cached()
-        return self.get_theory_vector_full(parameters)[mask]
-
-    def loglike(self, parameters):
-        diff = self.get_theory_vector_masked(parameters) - self.get_data_vector_masked()
-        inv_cov = self.get_covariance_matrix_masked_inv()
-        return -0.5 * diff @ inv_cov @ diff
+from cloelike.EuclidLikelihood_photo_base import PhotoLikelihoodBase
 
 
 class WLMixin:
