+++ cloelike/EuclidLikelihood_photo_Cls.py
@@ -1,143 +1,9 @@
 # file: euclid_likelihoods.py
 import numpy as np
-from copy import deepcopy
-from functools import lru_cache
 
 from cloelib.observables.photo import ShearTracer, PositionsTracer
 from cloelib.summary_statistics.angular_two_point import AngularTwoPoint
-
-
-class PhotoLikelihoodBase:
-    """
-    Base class for photometric likelihood calculations using angular power spectra (Cls).
-
-    Provides methods for preparing, binning, and masking data, and computing
-    likelihoods based on theoretical predictions and observed data vectors.
-
-    Parameters
-    ----------
-    data : dict
-        Input data required for likelihood computation.
-    settings : dict
-        Configuration settings for the likelihood calculation.
-    Background : object
-        Cosmological background instance.
-    LinPerturbations : object
-        Linear perturbations instance.
-    NonLinPerturbations : object
-        Non-linear perturbations instance.
-    mode : str, optional
-        Mode of operation, default is "coupled".
-    """
-
-    def __init__(
-        self,
-        data,
-        settings,
-        Background,
-        LinPerturbations,
-        NonLinPerturbations,
-        mode="coupled",
-    ):
-        self.data = data
-        self.settings = settings
-        self.derived = {}
-        self.Background = Background
-        self.LinPerturbations = LinPerturbations
-        self.NonLinPerturbations = NonLinPerturbations
-        self.mode = mode
-        self.scale_cuts = settings["scale_cuts"]
-        self.rebin = False
-        self.zs = data["z_arr"]
-        self.mixmat = deepcopy(data["mixmat"])
-        self._prepare()
-
-    # -------------------------------
-    #  Core preparation utilities
-    # -------------------------------
-    def _prepare(self):
-        """Perform rebinning if required by settings."""
-        if self.settings["n_ell_bins"] < len(self.data["ells"]):
-            self.rebin = True
-            self.data["cells_unbin"] = self.data["cells"]
-            self.data["ells_unbin"] = self.data["ells"]
-            self._bin_data(
-                self.data["cells"], self.data["ells"], self.settings["n_ell_bins"]
-            )
-            self._bin_mixmat()
-
-    def _bin_data(self, cells_data, ells, n_bins):
-        """Geometric rebinning of data vectors."""
-        bin_edges = np.geomspace(10, ells[-1], n_bins + 1)
-        mask_bins = [
-            (ells >= bin_edges[i]) & (ells < bin_edges[i + 1]) for i in range(n_bins)
-        ]
-        self.weight_mat = np.asarray(mask_bins, dtype=float)
-        self.weight_mat /= np.sum(self.weight_mat, axis=1)[:, None]
-        self.data["ells"] = np.array([np.mean(ells[mb]) for mb in mask_bins])
-        for k in cells_data.keys():
-            self.data["cells"][k] = cells_data[k] @ self.weight_mat.T
-
-    def _bin_mixmat(self):
-        """Apply the same binning to the mixing matrix."""
-        for k in self.mixmat.keys():
-            new_array = np.tensordot(self.weight_mat, self.mixmat[k], axes=([1], [-2]))
-            if k[:2] == ("SHE", "SHE"):
-                new_array = np.transpose(new_array, axes=(1, 0, 2))
-            self.mixmat[k] = new_array
-
-    def _masking(self, arr, interval):
-        """Helper: boolean mask for elements within a given interval."""
-        return (arr >= interval[0]) & (arr <= interval[1])
-
-    # -------------------------------
-    #  Default methods
-    # -------------------------------
-    def get_masking_vector(self):
-        """Return the combined masking vector (to be extended by mixins)."""
-        return np.array([], dtype=bool)
-
-    def get_covariance_matrix_full(self):
-        """Return the full covariance matrix."""
-        return self.data["cov"]
-
-    def get_data_vector_full(self):
-        """Return the full observed data vector (to be extended by mixins)."""
-        return np.array([])
-
-    def get_theory_vector_full(self, parameters):
-        """Return the full theoretical data vector (to be extended by mixins)."""
-        return np.array([])
-
-    @lru_cache(maxsize=None)
-    def get_masking_vector_cached(self):
-        """Compute (once) and cache the combined boolean mask."""
-        return self.get_masking_vector()
-
-    @lru_cache(maxsize=None)
-    def get_data_vector_masked(self):
-        """Compute (once) and cache masked data vector."""
-        mask = self.get_masking_vector_cached()
-        return self.get_data_vector_full()[mask]
-
-    @lru_cache(maxsize=None)
-    def get_covariance_matrix_masked_inv(self):
-        """Compute (once) and cache inverse masked covariance matrix."""
-        cov = self.get_covariance_matrix_full()
-        mask = self.get_masking_vector_cached()
-        cov_masked = cov[mask][:, mask]
-        return np.linalg.inv(cov_masked)
-
-    def get_theory_vector_masked(self, parameters):
-        """Return theoretical vector masked identically to the data vector."""
-        mask = self.get_masking_vector_cached()
-        return self.get_theory_vector_full(parameters)[mask]
-
-    def loglike(self, parameters):
-        """Compute Gaussian log-likelihood."""
-        diff = self.get_theory_vector_masked(parameters) - self.get_data_vector_masked()
-        inv_cov = self.get_covariance_matrix_masked_inv()
-        return -0.5 * diff @ inv_cov @ diff
+from cloelike.EuclidLikelihood_photo_base import PhotoLikelihoodBase
 
 
 class WLMixin:
