+++ cloelike/_version.py
@@ -28,7 +28,7 @@ version_tuple: VERSION_TUPLE
 commit_id: COMMIT_ID
 __commit_id__: COMMIT_ID
 
-__version__ = version = "0.2.2.dev9+g66d364853.d20251114"
-__version_tuple__ = version_tuple = (0, 2, 2, "dev9", "g66d364853.d20251114")
+__version__ = version = "0.2.2.dev10+gd40dada0a.d20251117"
+__version_tuple__ = version_tuple = (0, 2, 2, "dev10", "gd40dada0a.d20251117")
 
 __commit_id__ = commit_id = None
